requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv
pyyaml
httpx
beautifulsoup4
fastapi
uvicorn
rich
langchain
langchain-core
langchain-openai
langchain-anthropic
langchain-community
langgraph

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
bandit>=1.7.0
